LayoutTests/fast/forms/input-width-height-attributes-without-renderer-loaded-image.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script>
jsTestIsAsync = true;
</script>
</head>
<body>
<input type="image" id="imageElement" src="resources/green.jpg">
<div id="console"></div>
<script language="JavaScript">
    document.onreadystatechange = function() {
        readyState = document.readyState;
        if (readyState == 'complete') {
            e = document.getElementById("imageElement");
            e.style = "display:none";
            shouldBe('e.width', '16');
            shouldBe('e.height', '16');
            finishJSTest();
        }
    }
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
