safari-600.7-branch/LayoutTests/fast/forms/input-width-height-attributes-without-renderer-loaded-image-expected.txt


PASS e.width is 16
PASS e.height is 16
PASS successfullyParsed is true

TEST COMPLETE
